@@ -1,51 +1,26 @@
 {
-  "name": "client",
+  "name": "my-app",
   "version": "0.1.0",
   "private": true,
-  "homepage": "/user/login",
   "dependencies": {
-    "@date-io/date-fns": "^1.3.13",
-    "@emotion/cache": "^11.6.0",
-    "@emotion/react": "^11.7.0",
-    "@emotion/styled": "^11.6.0",
-    "@mui/icons-material": "^5.2.1",
-    "@mui/material": "^5.2.3",
-    "@mui/styles": "^5.2.3",
-    "@mui/x-date-pickers": "*",
-    "@testing-library/jest-dom": "^5.16.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "axios": "^0.24.0",
-    "date-fns": "^2.28.0",
-    "express": "^4.18.1",
-    "gematriya": "^2.0.0",
-    "moment": "^2.29.1",
-    "react": "^17.0.2",
-    "react-big-calendar": "^0.38.2",
-    "react-digraph": "*",
-    "react-dom": "^17.0.2",
-    "react-jewish-datepicker": "^1.2.1",
-    "react-router": "^6.1.0",
-    "react-router-dom": "^6.1.0",
-    "react-scripts": "4.0.3",
-    "react-spaces": "^0.3.2",
-    "react-transition-group": "^4.4.2",
-    "stylis": "^4.1.1",
-    "stylis-plugin-rtl": "^2.1.1",
-    "web-vitals": "^1.1.2"
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.3.0",
+    "@testing-library/user-event": "^13.5.0",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
+    "react-scripts": "5.0.1",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "cypress:open": "cypress open"
+    "eject": "react-scripts eject"
   },
   "eslintConfig": {
     "extends": [
       "react-app",
-      "react-app/jest",
-      "plugin:cypress/recommended"
+      "react-app/jest"
     ]
   },
   "browserslist": {
@@ -59,9 +34,5 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-  "devDependencies": {
-    "cypress": "^9.5.4",
-    "eslint-plugin-cypress": "^2.12.1"
   }
 }
